+++ src/main/g8/conf/server.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 
-# For Development Env       #
-# ------------------------- #
-# Start/Stop script on *NIX #
-# ------------------------- #
-# Command-line arguments:   #
-# -h help and exist         #
-# -a <listen-address>       #
-# -p <http-port>            #
-# -m <max-memory-in-mb>     #
-# -c <config-file.conf>     #
-# -l <logback-file.xml>     #
-# -j "extra-jvm-options"    #
-# ------------------------- #
+# For Non-Production Env                                                    #
+# ------------------------------------------------------------------------- #
+# Start/Stop script on *NIX                                                 #
+# ------------------------------------------------------------------------- #
+# Command-line arguments:                                                   #
+# -h help and exist                                                         #
+#    --pid <path-to-.pid-file>                                              #
+# -a|--addr <listen-address>                                                #
+# -p|--port <http-port>                                                     #
+# -m|--mem <max-memory-in-mb>                                               #
+# -c|--conf <path-to-config-file.conf>                                      #
+# -l|--logconf <path-to-logback-file.xml>                                   #
+#    --logdir <path-to-log-directory, $app.logdir will be set to this value #
+# -j|--jvm "extra-jvm-options"                                              #
+# ------------------------------------------------------------------------- #
 
 # from http://stackoverflow.com/questions/242538/unix-shell-script-find-out-which-directory-the-script-file-resides
 pushd \$(dirname "\${0}") > /dev/null
@@ -21,12 +23,11 @@ popd > /dev/null
 
 APP_HOME=\$_basedir/..
 APP_NAME=$name;format="normalize"$
-APP_PID=\$APP_HOME/\$APP_NAME.pid
 
 # Setup proxy if needed
 #APP_PROXY_HOST=host
 #APP_PROXY_PORT=port
-#APP_NOPROXY_HOST="localhost|127.0.0.1|10.*|192.168.*|host.domain.com"
+#APP_NOPROXY_HOST="localhost|127.0.0.1|10.*|192.168.*|*.local|host.domain.com"
 #APP_PROXY_USER=user
 #APP_PROXY_PASSWORD=password
 
@@ -35,12 +36,16 @@ DEFAULT_APP_PORT=9090
 DEFAULT_APP_MEM=64
 DEFAULT_APP_CONF=application.conf
 DEFAULT_APP_LOGBACK=logback-dev.xml
+DEFAULT_APP_PID=\$APP_HOME/\$APP_NAME.pid
+DEFAULT_APP_LOGDIR=\$APP_HOME/logs
 
 APP_ADDR=\$DEFAULT_APP_ADDR
 APP_PORT=\$DEFAULT_APP_PORT
 APP_MEM=\$DEFAULT_APP_MEM
 APP_CONF=\$DEFAULT_APP_CONF
 APP_LOGBACK=\$DEFAULT_APP_LOGBACK
+APP_PID=\$DEFAULT_APP_PID
+APP_LOGDIR=\$DEFAULT_APP_LOGDIR
 
 JVM_EXTRA_OPS=
 
@@ -72,7 +77,11 @@ doStop() {
 
 doStart() {
     echo -n "Starting \$APP_NAME: "
-    
+
+    if [ "\$APP_PID" == "" ]; then   
+        echo "Error: PID file not specified!"
+        exit 1
+    fi
     if [ -f "\$APP_PID" ]; then
         if isRunning \$APP_PID; then
             echo "Already running!"
@@ -82,7 +91,17 @@ doStart() {
             rm -f "\$APP_PID"
         fi
     fi
-    
+
+    if [ "\$APP_LOGDIR" == "" ]; then    
+        echo "Error: Log directory not specified!"
+        exit 1
+    else
+        mkdir -p \$APP_LOGDIR
+        if [ ! -d \$APP_LOGDIR ]; then
+            echo "Error: Log directory \$APP_LOGDIR cannot be created or not a writable directory!"
+        fi
+    fi
+
     if [ "\$APP_ADDR" == "" ]; then	
     	echo "Error: HTTP listen address not specified!"
         exit 1
@@ -127,7 +146,7 @@ doStart() {
         fi
     fi
     
-    RUN_CMD=(\$APP_HOME/bin/\$APP_NAME -Dapp.home=\$APP_HOME -Dhttp.port=\$APP_PORT -Dhttp.address=\$APP_ADDR)
+    RUN_CMD=(\$APP_HOME/bin/\$APP_NAME -Dapp.home=\$APP_HOME -Dapp.logdir=\$APP_LOGDIR -Dhttp.port=\$APP_PORT -Dhttp.address=\$APP_ADDR)
     RUN_CMD+=(-Dpidfile.path=\$APP_PID)
     if [ "\$APP_PROXY_HOST" != "" -a "\$APP_PROXY_PORT" != "" ]; then
         RUN_CMD+=(-Dhttp.proxyHost=\$APP_PROXY_HOST -Dhttp.proxyPort=\$APP_PROXY_PORT)
@@ -154,21 +173,25 @@ doStart() {
     echo "APP_MEM      : \$APP_MEM"
     echo "APP_CONF     : \$FINAL_APP_CONF"
     echo "APP_LOGBACK  : \$FINAL_APP_LOGBACK"
+    echo "APP_LOGDIR   : \$APP_LOGDIR"
     echo "APP_PID      : \$APP_PID"
     echo "JVM_EXTRA_OPS: \$JVM_EXTRA_OPS"
 }
 
 usageAndExit() {
-	echo "Usage: \${0##*/} <{start|stop}> [-h] [-m <memory limit in mb>] [-a <http listen address>] [-p <http listen port>] [-c <custom config file>] [-l <custom logback config>] [-j \"<extra jvm options>\"]"
-    echo "    stop : stop the server"
-    echo "    start: start the server"
-    echo "       -h : Display this help screen"
-    echo "       -m : JVM memory limit in mb (default \$DEFAULT_APP_MEM)"
-    echo "       -a : HTTP listen address (default \$DEFAULT_APP_ADDR)"
-    echo "       -p : HTTP listen port (default \$DEFAULT_APP_PORT)"
-    echo "       -c : Custom app config file, relative file is prefixed with ./conf (default \$DEFAULT_APP_CONF)"
-    echo "       -l : Custom logback config file, relative file is prefixed with ./conf (default \$DEFAULT_APP_LOGBACK)"
-    echo "       -j : Extra JVM options (example: \"-Djava.rmi.server.hostname=localhost)\""
+	echo "Usage: \${0##*/} <{start|stop|restart}> [-h] [--pid <.pid file>] [--logdir <log directory>] [-m <memory limit in mb>] [-a <http listen address>] [-p <http listen port>] [-c <custom config file>] [-l <custom logback config>] [-j \"<extra jvm options>\"]"
+    echo "    stop   : stop the server"
+    echo "    start  : start the server"
+    echo "    restart: restart the server"
+    echo "       -h or --help    : Display this help screen"
+    echo "       -m or --mem     : JVM memory limit in mb (default \$DEFAULT_APP_MEM)"
+    echo "       -a or --addr    : HTTP listen address (default \$DEFAULT_APP_ADDR)"
+    echo "       -p or --port    : HTTP listen port (default \$DEFAULT_APP_PORT)"
+    echo "       -c or --conf    : Custom app config file, relative file is prefixed with ./conf (default \$DEFAULT_APP_CONF)"
+    echo "       -l or --logconf : Custom logback config file, relative file is prefixed with ./conf (default \$DEFAULT_APP_LOGBACK)"
+    echo "       -j or --jvm     : Extra JVM options (example: \"-Djava.rmi.server.hostname=localhost)\""
+    echo "       --pid           : Specify application's .pid file (default \$DEFAULT_APP_PID)"
+    echo "       --logdir        : Specify application's log directory (default \$DEFAULT_APP_LOGDIR)"
     echo
     echo "Example: start server 64mb memory limit, with custom configuration file"
     echo "    \${0##*/} start -m 64 -c abc.conf"
@@ -190,6 +213,10 @@ while [ "\$1" != "" ]; do
     case \$PARAM in
         -h|--help)
             usageAndExit
+            ;;
+
+        --pid)
+            APP_PID=\$VALUE
             ;;
 
         -m|--mem)
@@ -216,11 +243,15 @@ while [ "\$1" != "" ]; do
             APP_CONF=\$VALUE
             ;;
             
-        -l|--log)
+        -l|--logconf)
             APP_LOGBACK=\$VALUE
             ;;
 
-        -j)
+        --logdir)
+            APP_LOGDIR=\$VALUE
+            ;;
+
+        -j|--jvm)
             JVM_EXTRA_OPS=\$VALUE
             ;;
 
@@ -237,6 +268,11 @@ case "\$ACTION" in
         ;;
 
     start)
+        doStart
+        ;;
+
+    restart)
+        doStop
         doStart
         ;;